rewrite SubSpriteMixin

+++ src/main/kotlin/com/axiom/axiom_pain/AxiomPainConfig.kt
@@ -11,7 +11,7 @@ object AxiomPainConfig {
 
     // This will hold the actual list of strings
     val ROBOT_USERNAMES: ForgeConfigSpec.ConfigValue<MutableList<out String>>
-    val TEMPERATURE_SHIFT: ForgeConfigSpec.ConfigValue<out Long>
+    val TEMPERATURE_SHIFT: ForgeConfigSpec.ConfigValue<out Int>
 
     init {
         BUILDER.push("Robot Settings")
@@ -22,7 +22,7 @@ object AxiomPainConfig {
 
         TEMPERATURE_SHIFT = BUILDER
             .comment("Long which is subtracted from the temperature. A value of 10000000 is roughly equivalent to a 0.1 float shift.")
-            .define("temperatureShift", 25000000L)
+            .define("temperatureShift", 25000000)
 
         BUILDER.pop()
         SPEC = BUILDER.build()
